[UPD] factorizacion de resources y repository para mayor legibilidad
