+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.application.name=indecBack
-server.servlet.context-path=/api/v1
+server.servlet.context-path=/api/v1/indec
 server.port=8084
 spring.datasource.url=jdbc:sqlserver://localhost:1433;databaseName=indec;encrypt=false
 spring.datasource.username=sa
